move the logic for build options to a common place to use it everywhere
